requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
flatbuffers==1.12
gast==0.3.3
grpcio==1.32.0
h11==0.12.0
h5py==2.10.0
idna==2.10
Markdown==3.3.4
# oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.1.1
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.1
six==1.15.0
starlette==0.13.6
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.4
Werkzeug==1.0.1
wrapt==1.12.1
pandas==1.2.3
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-hub==0.11.0
tensorflow-text==2.4.3
numpy==1.19.5
Keras==2.4.3
Keras-Preprocessing==1.1.2
opencv-python==4.5.1.48
